requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.23
pandas>=2.0
scikit-learn>=1.2
Pillow>=10.0
tqdm>=4.66
PyYAML>=6.0
